+++ background.ts
@@ -1,2 +1,3 @@
 import "./background/background-store"
 import "./background/icon-theme"
+import "./background/alarm-sound"