+++ themes/index.ts
@@ -1,4 +1,5 @@
 import type { PopupTheme } from "~popup/theme"
+import ghost from "./ghost.json"
 import dark from "./dark.json"
 import aggie from "./aggie.json"
 import light from "./light.json"
@@ -6,9 +7,10 @@ import green from "./green.json"
 
 // Add your themes here!!
 const allThemes: PopupTheme[] = [
-  light as PopupTheme,
+  ghost as PopupTheme,
   dark as PopupTheme,
   aggie as PopupTheme,
+  light as PopupTheme,
   green as PopupTheme,
 ]