+++ utils/debug-options.ts
@@ -1,4 +1,4 @@
-const DEBUG_MODE = false;
+const DEBUG_MODE = true;
 //console.log("blah")
 export function debugLog(text: string) {
   if (!DEBUG_MODE) { return; }